# improve sqllexer and obfuscator operation and memory efficiency