--- a/sqllexer.go
+++ b/sqllexer.go
@@ -92,7 +92,8 @@ func (s *Lexer) Scan() Token {
 	case isLeadingSign(ch):
 		// if the leading sign is followed by a digit, then it's a number
 		// although this is not strictly true, it's good enough for our purposes
-		if isDigit(s.lookAhead(1)) || s.lookAhead(1) == '.' {
+		nextCh := s.lookAhead(1)
+		if isDigit(nextCh) || nextCh == '.' {
 			return s.scanNumber(true)
 		}
 		return s.scanOperator()
@@ -104,11 +105,6 @@ func (s *Lexer) Scan() Token {
 		if isDigit(s.lookAhead(1)) {
 			// if the dollar sign is followed by a digit, then it's a numbered parameter
 			return s.scanNumberedParameter()
-		}
-		if s.matchAt([]rune("$func$")) {
-			// check for dollar quoted function
-			// dollar quoted function will be obfuscated as a SQL query
-			return s.scanDollarQuotedFunction()
 		}
 		return s.scanDollarQuotedString()
 	case isOperator(ch):
@@ -185,8 +181,7 @@ func (s *Lexer) scanNumber(leadingSign bool) Token {
 }
 
 func (s *Lexer) scanDecimalNumber() Token {
-	s.next()
-	ch := s.peek()
+	ch := s.next()
 
 	// scan digits
 	for isDigit(ch) || ch == '.' || isExpontent(ch) {
@@ -224,28 +219,26 @@ func (s *Lexer) scanOctalNumber() Token {
 
 func (s *Lexer) scanString() Token {
 	s.start = s.cursor
-	s.next() // consume the opening quote
+	ch := s.next() // consume the opening quote
 	escaped := false
 
 	for {
-		ch := s.peek()
-
 		if escaped {
 			// encountered an escape character
 			// reset the escaped flag and continue
 			escaped = false
-			s.next()
+			ch = s.next()
 			continue
 		}
 
 		if ch == '\\' {
 			escaped = true
-			s.next()
+			ch = s.next()
 			continue
 		}
 
 		if ch == '\'' {
-			s.next() // consume the closing quote
+			ch = s.next() // consume the closing quote
 			return Token{STRING, s.src[s.start:s.cursor]}
 		}
 
@@ -254,14 +247,14 @@ func (s *Lexer) scanString() Token {
 			// this usually happens when the string is truncated
 			return Token{INCOMPLETE_STRING, s.src[s.start:s.cursor]}
 		}
-		s.next()
+		ch = s.next()
 	}
 }
 
 func (s *Lexer) scanIdentifier() Token {
 	// NOTE: this func does not distinguish between SQL keywords and identifiers
 	s.start = s.cursor
-	ch := s.peek()
+	ch := s.next()
 	for isLetter(ch) || isDigit(ch) || ch == '.' || ch == '?' {
 		ch = s.next()
 	}
@@ -271,15 +264,14 @@ func (s *Lexer) scanIdentifier() Token {
 
 func (s *Lexer) scanDoubleQuotedIdentifier() Token {
 	s.start = s.cursor
-	s.next() // consume the opening quote
+	ch := s.next() // consume the opening quote
 	for {
-		ch := s.peek()
 		// encountered the closing quote
 		// BUT if it's followed by .", then we should keep going
 		// e.g. postgre "foo"."bar"
 		if ch == '"' {
 			if s.matchAt([]rune(`"."`)) {
-				s.nextBy(3) // consume the "."
+				ch = s.nextBy(3) // consume the "."
 				continue
 			}
 			break
@@ -287,7 +279,7 @@ func (s *Lexer) scanDoubleQuotedIdentifier() Token {
 		if isEOF(ch) {
 			return Token{ERROR, s.src[s.start:s.cursor]}
 		}
-		s.next()
+		ch = s.next()
 	}
 	s.next() // consume the closing quote
 	return Token{IDENT, s.src[s.start:s.cursor]}
@@ -296,7 +288,7 @@ func (s *Lexer) scanDoubleQuotedIdentifier() Token {
 func (s *Lexer) scanWhitespace() Token {
 	// scan whitespace, tab, newline, carriage return
 	s.start = s.cursor
-	ch := s.peek()
+	ch := s.next()
 	for isWhitespace(ch) {
 		ch = s.next()
 	}
@@ -305,7 +297,7 @@ func (s *Lexer) scanWhitespace() Token {
 
 func (s *Lexer) scanOperator() Token {
 	s.start = s.cursor
-	ch := s.peek()
+	ch := s.next()
 	for isOperator(ch) {
 		ch = s.next()
 	}
@@ -320,8 +312,7 @@ func (s *Lexer) scanWildcard() Token {
 
 func (s *Lexer) scanSingleLineComment() Token {
 	s.start = s.cursor
-	s.nextBy(2) // consume the opening dashes
-	ch := s.peek()
+	ch := s.nextBy(2) // consume the opening dashes
 	for ch != '\n' && !isEOF(ch) {
 		ch = s.next()
 	}
@@ -330,11 +321,10 @@ func (s *Lexer) scanSingleLineComment() Token {
 
 func (s *Lexer) scanMultiLineComment() Token {
 	s.start = s.cursor
-	s.nextBy(2) // consume the opening slash and asterisk
+	ch := s.nextBy(2) // consume the opening slash and asterisk
 	for {
-		ch := s.peek()
 		if ch == '*' && s.lookAhead(1) == '/' {
-			s.nextBy(2) // consume the closing asterisk and slash
+			ch = s.nextBy(2) // consume the closing asterisk and slash
 			break
 		}
 		if isEOF(ch) {
@@ -342,7 +332,7 @@ func (s *Lexer) scanMultiLineComment() Token {
 			// this usually happens when the comment is truncated
 			return Token{ERROR, s.src[s.start:s.cursor]}
 		}
-		s.next()
+		ch = s.next()
 	}
 	return Token{MULTILINE_COMMENT, s.src[s.start:s.cursor]}
 }
@@ -353,58 +343,38 @@ func (s *Lexer) scanPunctuation() Token {
 	return Token{PUNCTUATION, s.src[s.start:s.cursor]}
 }
 
-func (s *Lexer) scanDollarQuotedFunction() Token {
-	s.start = s.cursor
-	s.nextBy(6) // consume the opening dollar and the function name
-	for {
-		ch := s.peek()
-		if ch == '$' && s.matchAt([]rune("$func$")) {
-			break
-		}
-		if isEOF(ch) {
-			return Token{ERROR, s.src[s.start:s.cursor]}
-		}
-		s.next()
-	}
-	s.nextBy(6) // consume the closing dollar quoted function
-	return Token{DOLLAR_QUOTED_FUNCTION, s.src[s.start:s.cursor]}
-}
-
 func (s *Lexer) scanDollarQuotedString() Token {
 	s.start = s.cursor
-	s.next() // consume the dollar sign
-	dollars := 1
-	for {
-		ch := s.peek()
-		if ch == '$' {
-			// keep track of the number of dollar signs we've seen
-			// a valid dollar quoted string is either $$string$$ or $tag$string$tag$
-			// we technically should see 4 dollar signs
-			// this is a bit of a hack because we don't want to keep track of the tag
-			// but it's good enough for our purposes of tokenization and obfuscation
-			dollars += 1
-		}
-		if isEOF(ch) {
-			break
-		}
-		s.next()
-	}
+	ch := s.next() // consume the dollar sign
+	tagStart := s.cursor
 
-	if dollars != 4 {
-		return Token{UNKNOWN, s.src[s.start:s.cursor]}
+	for s.cursor < len(s.src) && ch != '$' {
+		ch = s.next()
 	}
-	return Token{DOLLAR_QUOTED_STRING, s.src[s.start:s.cursor]}
+	ch = s.next()                       // consume the closing dollar sign of the tag
+	tag := s.src[tagStart-1 : s.cursor] // include the opening and closing dollar sign e.g. $tag$
+
+	for s.cursor < len(s.src) {
+		if s.matchAt([]rune(tag)) {
+			s.nextBy(len(tag)) // consume the closing tag
+			if tag == "$func$" {
+				return Token{DOLLAR_QUOTED_FUNCTION, s.src[s.start:s.cursor]}
+			}
+			return Token{DOLLAR_QUOTED_STRING, s.src[s.start:s.cursor]}
+		}
+		ch = s.next()
+	}
+	return Token{ERROR, s.src[s.start:s.cursor]}
 }
 
 func (s *Lexer) scanNumberedParameter() Token {
 	s.start = s.cursor
-	s.nextBy(2) // consume the dollar sign and the number
+	ch := s.nextBy(2) // consume the dollar sign and the number
 	for {
-		ch := s.peek()
 		if !isDigit(ch) {
 			break
 		}
-		s.next()
+		ch = s.next()
 	}
 	return Token{NUMBERED_PARAMETER, s.src[s.start:s.cursor]}
 }
@@ -414,11 +384,4 @@ func (s *Lexer) scanUnknown() Token {
 	s.start = s.cursor
 	s.next()
 	return Token{UNKNOWN, s.src[s.start:s.cursor]}
-}
-
-func (s *Lexer) skipWhitespace() {
-	ch := s.peek()
-	for isWhitespace(ch) {
-		ch = s.next()
-	}
 }