… obfuscator_test.go
… sqllexer_bench_test.go
… sqllexer_test.go