This PR greatly improves the sqllexer and obfuscator operation and memory efficiency with changes
- Reduce repeated lookahead by skipping checks for already scanned characters
- Completely eliminate runes to string conversion and runes to string comparison
- Removed regex based obfuscation
- Obfuscator switched to `ScanAll`, which pre-scans all tokens. The prior implementation uses`ScanAllToken` that creates a goroutine and emits tokens to the channel. See benchmark results below.

Some housekeeping changes
- Add benchmark for sqllexer
- Fix the ordering of expected vs. actual in `assert.Equal`

Obfuscator Benchmark before vs. after
- Before (scans token on the fly)
```
pkg: github.com/DataDog/go-sqllexer
BenchmarkObfuscator/Escaping/512-10         	   59818	     25308 ns/op	    2480 B/op	      85 allocs/op
BenchmarkObfuscator/Grouping/199-10         	   67796	     18358 ns/op	    1864 B/op	      69 allocs/op
BenchmarkObfuscator/Large/3694-10           	    5238	    274721 ns/op	   29544 B/op	     887 allocs/op
BenchmarkObfuscator/Complex/969-10          	   16414	     72819 ns/op	    9256 B/op	     253 allocs/op
BenchmarkObfuscator/SuperLarge/4198-10      	    4518	    246665 ns/op	   38752 B/op	     878 allocs/op
```

- After (pre-scan all tokens)
```
pkg: github.com/DataDog/go-sqllexer
BenchmarkObfuscator/Escaping/512-10         	  282336	      4441 ns/op	    6624 B/op	      14 allocs/op
BenchmarkObfuscator/Grouping/199-10         	  489855	      2407 ns/op	    3296 B/op	      12 allocs/op
BenchmarkObfuscator/Large/3694-10           	   26586	     48246 ns/op	   86240 B/op	      23 allocs/op
BenchmarkObfuscator/Complex/969-10          	  111474	     11046 ns/op	   15584 B/op	      18 allocs/op
BenchmarkObfuscator/SuperLarge/4198-10      	   21439	     55041 ns/op	   95712 B/op	      25 allocs/op
```

Generator pattern, often more memory efficient for large data sets, does introduce overhead due to channel communication and goroutine management. Pre-scanning all tokens in this case are much faster.